KB simplify version
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@52604 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/service/com/liferay/knowledgebase/model/ArticleConstants.java
@@ -17,10 +17,11 @@ package com.liferay.knowledgebase.model;
 /**
  * <a href="ArticleConstants.java.html"><b><i>View Source</i></b></a>
  *
+ * @author Peter Shin
  * @author Brian Wing Shun Chan
  */
 public class ArticleConstants {
 
-	public static final double DEFAULT_VERSION = 1.0;
+	public static final int DEFAULT_VERSION = 1;
 
 }
